some updates
A new program is added : KocksMecking2018. In this one the only change
from TensileTestAnalysis code is that in calculating KM curve. In this
one the True Stress-Strain  curve is not fitted by a polynomial, instead
there are four different methods. It can chose peaks, valleys, n random
points or do a moving average of n points along the curve.